Move interfaces under a common namespace
