+++ tests/FacadeTest.php
@@ -2,8 +2,8 @@
 
 namespace hisorange\BrowserDetect\Test;
 
+use hisorange\BrowserDetect\Contracts\ParserInterface;
 use hisorange\BrowserDetect\Facade;
-use hisorange\BrowserDetect\ParserInterface;
 
 /**
  * Class FacadeTest
